refactor: feedback from pr
